backend/requirements.txt
django==5.2.4
djangorestframework==3.14.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
django-cors-headers==4.3.1